01-Code/backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# File parsing
pdfplumber==0.10.3
python-docx==1.1.0
PyMuPDF==1.23.21

# Rate limiting
slowapi==0.1.9

# AI/ML
openai==1.12.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
httpx==0.26.0

# Future: AI/ML (when implemented)
# openai==1.10.0
# langchain==0.1.0
